MGDCTRS-699 chore: refactor the i18n json
This PR updates the i18n json file to use a consistent camel case
convention for i18n keys.  It also sorts the json by key and removes a
number of duplicate entries.  This also updates most of the strings as
requested in MGDCTRS-699 that are in the i18n configuration, follow up
commits will be needed for components that haven't yet been localized.
Given all the changes this commit also removes the Italian translation,
we can create a new one once things are more finalized.