fix MGDCTRS-800, MGDCTRS-759 (#451)
This change overhauls the way JSON schema driven forms are rendered,
leaning more on the functionality available from the custom JSON schema
bridge.  This change also adds a simple $ref lookup function, a simple
object from schema creator and a couple quick tests for each.  Also this
change refactors the clearSecretEmptyValue function, it now returns a
new object with the changed values, and adds a test for that as well.