MGDCTRS758 fix: make credential fields editable
This change changes the edit and duplicate page flows so that users can
edit secret fields such as secret keys and client IDs that are part of
JSON schema driven forms.  This commit also adds some text to unset
fields to clearly indicate to the user that a configuration field is
unset.  Also this change ensures that configuration that doesn't include
a data shape is catered for client-side in both the duplicate and edit
cases.  And finally, this commit also adds some help text to inform the
user why a credential field's value isn't shown.