chore: get react refresh working
This change adds the necessary configuration for react refresh and
ts-loader.  It also removes the extra entrypoint that was causing two
client side instances of the hot module loading code to run.  This
ensures that code changes will be reloaded in the browser properly.
Also this commit favors the error overlay from react refresh over the
built in one from webpack dev server.