MGDCTRS-1044 fix: ensure unset enums show as unset (#495)
This change addresses how uniforms can represent what is valid data in
an inconsistent way.  For enums if the underlying model object is unset,
uniforms will still show the first enum value, leading the user to
believe that the field is configured.  This change works around this by
inserting empty enum values as needed into the schema definition so it's
not necessary for the schema to have to deal with a UI concern.