MGDCTRS-968 fix: show all the connectors
This change is intended to temporarily remove the pagination control for
the connector types step in the connector creation wizard so that all
connector types are visible to the user when the step loads.  It also
adds an argument to the pagination machine so that a given view can
control the default page size.