MGDCTRS-772 feat: server-side filtering/ordering
This change updates the pagination utility classes to add support for
orderBy and aligns the filtering support to better match the API.  This
change also updates the pagination usage a little bit across the
toolbars in the application.  Finally this change also brings back the
search filter on the connectors instance list.  Further work is needed
to add sorting controls as well as additional filter categories for the
connectors instance list search.