MGDCTRS-772 feat: server-side filtering/ordering (#513)
* MGDCTRS-772 feat: server-side filtering/ordering

This change updates the pagination utility classes to add support for
orderBy and aligns the filtering support to better match the API.  This
change also updates the pagination usage a little bit across the
toolbars in the application.  Finally this change also brings back the
search filter on the connectors instance list.  Further work is needed
to add sorting controls as well as additional filter categories for the
connectors instance list search.

* MGDCTRS-968 feat: show more connector types

This change bumps the number of connector types shown by default to 20
and also adds a pagination control underneath the gallery element so the
user can see that there's more pages to select from.

* MGDCTRS-772 fix: ensure string queries are quoted

This change wraps search queries in quotes to ensure spaces are
preserved properly in the resulting SQL statement.