MGDCTRS-862 fix: send a null to clear empty values
This change adjusts the form to produce a null value for unset fields as
well as update the field validation to allow for this.  This change also
adds an integration test that starts to cover the overview page as well
as an integration test that validates an unset field is sent as null.
Some additional enhancements to the integration tests are more
consistent naming and the test URLs have all been adjusted so it's clear
that the tests are not invoking actual URLs.  One final change is that
the Kafka management API URI that's set at the root of the app is passed
along through the create connector wizard machine.