MGDCTRS-1068 feat: state driven user analytics
This change adjusts the wizard state machine to emit activity events as
the user proceeds through the wizard.  These activity events are
funneled through a sendAnalytics action which will (once connected) post
user events.  This change also aligns the machine names and IDs to match
up with the current UI.