MGDCTRS-1068 feat: interpret activity from state
This change adds an interpreter to the wizard state machine that
converts state transitions to user activity events.  Currently these
events are directly connected to a tracking API when the app is used in
federated mode, in demo and e2e test mode these activity events are
logged to debug.