MGDCTRS-1542 fix: avoid extra refreshes
This change fixes an issue where the core configuration step generates
more machine events than are necessary and cleans up a few warnings in
the javascript console by supplying default values to the fields that
are rendered in the UI.  This change also continues to update some of
the names to be better in sync with the UI.  In addition, a suggested
configuration value is applied to all of the xstate machines to further
clean up warnings that are currently logged to the javascript console in
development mode.  This also includes an update to take care of a
deprecation warning for MSW as well in storybook.  Finally, this change
also reworks how the state machine's guards are set up and ensures that
they're used to check the validity of the connector configuration which
now is what will enable/disable the final finish/save button in the
wizard.