MGDCTRS-1068 fix: add an analytics hook
This change updates how analytics are made available to the app via
using a custom hook that's configurable via a context.  This change also
updates the directory structure and filenames for clarity, i.e.
"context" to "hooks" for hooks that cover app-wide concerns such as
analytics or basic app configuration like the existing useCos hook.