MGDCTRS-1551 fix: restructure overview page
This change restructures the connector overview page, moving the data
fetching logic out of the page component and into a context that just
exposes the fetched data or any error that results in fetching the
connector.  This change also adds an empty state with a back button
should the UI fail to fetch the connector on a poll to cater for unknown
errors and especially for cases where a user may have the overview page
open when that connector instance is deleted.  This change also adds a
story for the overview page that mocks all of the API used by the page
and further moves around some of the existing stories to tidy up the
tree of stories a bit.