MGDCTRS-1446 chore: update to the latest TableView
This change updates the Connector instances page to use the latest
version of the TableView component.  The TableView component now
includes the toolbar functionality which required some adaptation to
maintain the same filtering ability.  A further improvement however
would be to add the ability to filter on more connector fields, and to
update the filtering and pagination to use URL search parameters.  Some
other changes include exposing the selected state machine service and a
general restructuring of the instances table code to ensure the loading
is consistent.  This change also adds an additional story so it's
possible to see how the pagination controls behave.