MGDCTRS-1477 feat: categorized gallery spike
This commit adds an initial spike of the categorized connector gallery
view into storybook.  The storybook includes enough of a backend
implementation to provide live data to the connector type gallery view
so that the filtering controls work as expected.  Some adjustments to
the names of some types within the app are included with this commit,
though most additions are under the ProofOfConcepts directory.