MGDCTRS-1438 feat: dynamic form improvements
This change adds a method to start applying some general organization to
the schema driven forms.  The kafka topic should be moved to the start
of the form, and for AWS based connectors the access key/secret fields
are next, followed by the region.  The other change this commit brings
in is a new custom control for schema based forms to improve the AWS
region selector so that it behaves more like a typeahead control and
less like a select control.