MGDCTRS-1477 feat: improve loading cards
This change swaps out the loading spinner for a card containing skeleton
elements, and ensures that for either the list view or the card view
that the card sizes are at fixed sizes to optimize the viewport
scrolling.  Also this change brings in the latest patternfly version to
gain access to the new source/sink icons.