MGDCTRS-1619 feat: show inline validation errors
This change connects the validator to uniforms so that validation issues
are shown to the user.  For now this is only done for fields that the
user has interacted with to avoid highlighting all empty fields.  This
commit also adds a story and fixture for the http-sink connector, as it
uses the "pattern" property to constrain the acceptable input.  This
change adds customization and filtering functions to control what errors
are shown to the user and deal with some edge cases.