MGDCTRS-1832 fix: update the label querying
This change updates the way the UI builds the label query.  It also
updates the psuedo backend implementation to support the new query
syntax.  This commit also syncs up the query built for the name and
pricing tier to be consistent.  When querying for labels the UI will
avoid using the aggregate query if the user has only selected one
filter.  The function used to fetch the labels has been migrated to use
the API generated function.