MGDCTRS-1850 chore: add a link to the SA guide
This commit adds some text and a link to the core configuration step to
the user guide that covers service account creation and configuration
for Kafka.  This change also adds reworks the core configuration step
and adds some stories to storybook so it's easier to develop this step
outside of the application.  This commit modifies the core configuration
step so that the user needs to check a checkbox to confirm that the
service account configured has been set up to access the Kafka
instance.  The checkbox is unchecked when creating a connector, but
checked when duplicating a connector.

This commit also adjusts how the top level categories are shown in
storybook to ideally make it easier to find the right component.