MGDCTRS-1852 feat: show the selected connector
This commit reworks how the wizard header is defined, removing it from
the page section outside of the wizard and adopting the WizardHeader
component with several overrides.  This commit aligns the create and
duplicate connector pages to use the same header component.  This change
also removes some previously unused code from UncontrolledWizard and
updates it to accept an external header component.  Finally, this change
also updates the integration test to look for the selected connector
name once it's been clicked in the wizard.