MGDCTRS-1773 feat: better connector error display
This commit adopts the new inline error alert components into the app in
both the connector drawer and the connector overview page.  This change
also promotes these components from the ProofOfConcepts directory and
renames the components to closer match their usage.  Also this change
contains a re-alphabetize of the locales file.