MGDCTRS-1912 fix: remove sort control
This change removes the sort control from the connector type selector
toolbar to save some screen real estate.  This commit also updates icon
used when the filter control collapses in narrower viewports.  Finally,
this change also includes a small tweak to the order in the labels for
featured items to be consistent.