MGDCTRS-1926 chore: migrate useAlert usage
This change introduces a local useAlert hook to replace the
implementation from app-services-ui-shared.  In federated mode this hook
is now connected to the console notification system via
frontend-components-notifications, and this component is now a direct
dependency.  This commit therefore also contains some required
initialization code and dependencies such as a redux store and
react-redux for frontend-components-notifications.  Finally, this commit
also changes the routing behavior when saving a configuration change,
the user remains in the connector detail/overview page when editing a
connector instead of going back to the connector list.