MGDCTRS-1926 chore: remove useConfig usage
This change removes usage of useConfig from app-services-ui-shared.  The
data it was providing has been moved to the useCos hook.  For the
federated entrypoint AppFederated.tsx the mapping that exists in
app-services-ui is now local and added to a new endpoints file as
constant values.  The federated entrypoint will consult this map during
initialization and will use the appropriate backend based on the
hostname used to access the console.  The standalone (AppDemo) and e2e
(AppE2E) entrypoints behave as before but each contain their own
specific setup to cater for their usage.