MGDCTRS-2224 feat: create ErrorHandlerInfo
This change extracts out the logic that displays a connector's error
handler info into it's own component and updates it to provide a
DescriptionList variant.  This will now be visible in the drawer on the
connector instances page.  The existing component on the instance
configuration tab uses this updated component as well in form view.
This commit also reworks some of the page structure for the instances
page to better accept these changes as well as additional changes that
need to be made in this drawer.

The connector instances page story has been enhanced to also generate
dummy responses to fetching kafka instances and connector namespaces.

This change also adds the error handling card to the connector overview
page and updates the layout to match the design.